refactor: split edge cache contract from policy
